Navigate documents using XML namespaces where possible
A static dictionary is created mapping names to a number of well-known
namespaces. Use the names later on where possible, which reduces the
need for horrible constructs like local-name().

One exception I found is that the file we look in for Kobo-specific DRM
doesn't use namespaces.

I tested this with a number of books (using no DRM, Adobe, Apple and
Kobo DRM) and they all seemed to have appropriate levels of detail
extracted.