Replace author string with author array
This requires the generator needs to insert the commas instead of the
JTPepub class.

The authors method is built using creatorsWithOPFRole, modelled on
contributorsWithOPFRole. These two could probably use common code.

A test for a title containing funny XML characters is added. The XPath
query should be returning unescaped strings, this makes sure.

The translator and illustrator tests are corrected.

The test for three authors is corrected to check each value.