Add test files with fake DRM
The three files are fake - built with Sigil, unzipped, the extra file
inserted from an actual book with the right flavour DRM, and then zipped
up again. The results are only enough for JTPepub to recognize the
scheme, and not really DRM protected.

Tests are added to check the schemes are detected, also that no DRM is
detected in the Untitled file.