Swap colours in leopard.css
The colours for td and th look better the other way around, so the
values (the interesting bits) are brighter.

I also changed the colours so that the same "dimmed" colour used by
the author was used for the th.

It looks OK in Snow Leopard and I can see metadata properly.