Detect B&N DRM
Barnes & Noble use a variant of Adept DRM. One obvious difference seems
to be that contents of <operatorURL> contains a barnesandnoble.com URL.

This change switches from looking at <licenseURL> to <operatorURL> and
checking the URL contents.

The adeptXML object cannot be released if we want to look inside a
returned node, so we now release that later on.

The existing tests still pass.