Revert the change for finding creators
The opf:role attribute is optional, and I've seen books without it on
dc:creator. This change restores the previous code which matches those
role-less creators.

It does that in a surprising way - itemID is nil for those elements, and
it turns out the caseInsensitiveCompare: method returns NSOrderedSame
for that.