Detect and report expiring library books
The only library books I have are from Overdrive (via the London
Libraries Consortium), which use Adept DRM. The rights.xml file contains
<until> elements.

So while parsing Adept DRM, we look for <until> elements and build a
date if possible.

The parsing of the date string is moved to a category on NSDate; the
parsing of dates in ePub is reasonably well defined and it is not what
Cocoa call "natural language". I switched the publication date parsing
to using this category as well.

It is not known if the DRM expiry dates are formatted the same as dates
elsewhere in ePubs.

The expiry date is rendered into the HTML.

Tests of the date parser are added. Tests of a fake library book are
added as well.