Use mmap-based I/O instead of stdio
This is a little faster on small epubs. On bigger files - 70MB, 300MB -
this is about twice the speed. File sizes seem to be increasing with
lots of multimedia in ebooks, so this is probably a useful change.

I benchmarked this using "qlmanage -p -z" on a number of files.

A major omission from this code is that writing is not supported.

The unit tests still pass.