Fix leak, build in category, report DRM
The JTPepub object was leaked in the importer.

The duedate stuff wasn't working because the NSDate+OPF category needs
to be built in - this was apparent when doing an mdimport -V of a folder
containing a library book.

Reported the DRM scheme in the kMDItemSecurityMethod attribute. That's
currently used for PDF files and sort of fits for this. One oddity is
that PDF files set this attribute to "None", so we do that too.

I tried to clarify the expected data types for each attribute.