Rewrite -creators
This now finds all the creators *and* all the contributors and returns
an array with just their names instead of names and roles.

I refactored the logic to extract names from file-as attributes into a
new method.

Some test cases are added which check creators returns all the expected
values for the Untitled and metadata books.

The test cases all succeed.