Store synopis in kMDItemHeadline
The wrinkle here is that kMDItemHeadline is plain text, and the synopsis
might be HTML. This code removes anything that looks like an HTML tag
from the returned string.

Some unit tests probably ought to be added for this. It works on a few
files with HTML-ish dc:descriptions, and a few files with plain text
dc:descriptions.