Fix ISBN detection heuristics and include it in Spotlight
The ISBN detector was going wrong on (dubious) input like this:

    <dc:identifier>999</dc:identifier>
    <dc:identifier opf:scheme="ISBN">1234567</dc:identifier>

This effectively did:

if ([nil caseInsensitiveCompare:@"ISBN"] == NSOrderedSame) {

And that was true, so "999" got returned! Switching nil and @"ISBN"
around works better and is logically identical.

An even better fix might be to use the opf:scheme value if found and if
not use the one without an opf:scheme. But that might be going too far,
and I'm not sure what the correct semantics for dc:identifier really
are.

I've added the ISBN to Spotlight as the kMDItemIdentifier property.