Construct mimetype string safely
The stringWithUTF8String method assumes the pointer being passed in is
to a string of NUL terminated UTF-8 characters. It wasn't though, the
ZipArchive code was only mallocing sufficient bytes for the actual
characters.

This caused much strangeness depending on what memory happened to be
right after the malloced data. Symptoms included some tests failing, the
mimetype being set to nil, the mimetype being set to a longer string,
QuickLook not using our plugin for some files and then using it the next
time...

This addresses issue 24.