Clean up ZipArchive initializers
The init method was removed, as it wasn't being used and wasn't useful.
The initWithZipFile method now does the real work.

The initWithZipFile method is fixed to not dealloc itself.

The dataForNamedFile method is altered a little to avoid a memory
leak of 'buffer' on decompression failure.

Unit tests were run repeatedly and still pass.