Code cleanup
BOOL variables should use YES/NO symbols, not true/false etc.

The container object is autoreleased, so sending release is an error.

The content object is autoreleased so an extra retain/release is
unnecessary.

The cover object is retained already, so another retain will cause it to
be leaked.

The coverData object is autoreleased, so sending release is an error.

The retain of publicationDate is moved out of the loop, in case it gets
set multiple times in the loop. That would cause previous values to
leak.

The unit tests still pass.