Convert XHTML items in manifest into plain text
The Spotlight importer is going to use this, but doesn't yet.

The model is that it requests items (0..) from the manifest, which are
plaintext (not HTML) NSStrings. The manifest is lazily constructed by
searching the OPF for items with the right mime type. Getting nil back
means you've completed the manifest.

I'm not sure what it will return for encrypted books.

Conversion to plaintext is done using a streaming XML parser. I started
using the DOM parser, but that doesn't handle named entities like &foo;.
DOM parsers also use a lot of memory.

So I used Apple's NSXMLParser, which is a SAX streaming parser that uses
a lot less memory. It calls a delegate (set here to self) when element
tags are read and when characters are read. The text is accumulated into
JTPepub's capturing ivar.

The wrinkle is that it doesn't handle named entities either. But it is
easy to support them - see the
parser:resolveExternalEntityName:systemID: method. This lazily loads in
a dictionary of entity names to Unicode strings which I took from W3C.

http://www.w3.org/TR/xml-entity-names/xhtml1-lat1.html
http://www.w3.org/2003/entities/2007/xhtml1-special.ent
http://www.w3.org/2003/entities/2007/xhtml1-symbol.ent

Unit tests pass. I added another one which reads all the chapters from
the Untitled epub, which I modified to have 3 chapters with some
character entities.