Index book content for Spotlight
If there's no DRM, iterate through the manifest XHTML items and
accumulate the text contents. Add the text contents to
kMDItemTextContent.

For debugging, I've added the number of characters indexed to the
kMDItemComment. You can see this field in mdls and it gives a clue that
the book's been indexed.

After I use mdimport to rescan my ebooks (mds works in the background?),
I can mdfind text in them and use the Spotlight menu to find things.

Unit tests still pass.