Add a custom schema file
This has two benefits.

The first is that we can create custom attributes for things we can
extract (e.g. illustrators) that Spotlight has no concept of (i.e. the
kMDItemFoo things)

So this creates attributes for illustrators and translators. Note the
name is reverse DNS style with "." replaced with "_". Note also that "-"
needs to be replaced or the schema file will be unparseable at runtime -
I used "_" for that too.

The second benefit is that this allows us to customize the "More info"
panel in the Finder "Get Info" window. This is quite cool.

It would be nice if the More info panel could display our custom
attributes, but it doesn't seem to. There's no obvious way for it to map
the attribute name org_idpf_foo to a friendly label.

I commented out the "Indexed nnnn characters" things.