Display language in QL preview
The trick is how to get localized names for the language tags! The
NSLocale class can be initialized with one of the tags and can return
the localized value of the name.

Hopefully that means a book in "de" will be displayed as "German" for
an English user, and "Deutsch" for a German user.