href attributes need URI decoding
Two methods read "filenames" out of the OPF. These aren't filenames,
they're URIs and so need to be unescaped in order to be usable as
filenames in the zip.

NSString has a handy -stringByReplacingPercentEscapesUsingEncoding:
method which is used here.

The tests still all succeed. My book with a space in the cover image
filename now displays a cover image.